requirements.txt
google-api-python-client==2.90.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
selenium==4.10.0
webdriver-manager==3.8.6
